Implemented hotkey combinations! This closes #3.
This appears to be working fantastically. The one drawback, if you can
call it that, is that the meta keys (ctrl, alt, shift) cannot be used as
hotkeys themselves - only as parts of combinations of other hotkeys.
This may or may not need to be addressed/improved upon later in dev.
cycle.